Brings the post-2026-05-22 patch opcode rediscovery (everquest.h,
opcodes.toml + generated XMLs) and the map-package provider feature
into the multibox branch so a fresh multi-client capture decodes
against current Live wire format.

Conflict resolutions (all additive — both features coexist):
- sessionadapter.{h,cpp}: keep box-list (sendBoxList/onActiveBoxChanged,
  m_boxes) alongside map-package (sendMapPackages, m_mapPackageHost);
  both initial sends fire on Subscribe.
- wsserver.h: keep m_boxes/m_firstClientFired + m_mapPackageHost.
- daemonapp.cpp: setState() passes boxRegistry AND m_ws->setMapPackageHost(this).
- proto gitlink -> 0272d75 (proto main; unified box + map-package + lock).

+++ src/player.h
@@ -293,7 +293,7 @@ public:
   bool m_freshKill;
   
   // last spell cast on this player
-  uint16_t m_lastSpellOnId;
+  uint32_t m_lastSpellOnId;
   
   int16_t m_headingDegrees;
   // Wether or not we use defaults, determined by wether or not we could
