+++ src/util.h
@@ -51,7 +51,7 @@ char *print_addr (unsigned long addr);
 QString Commanate (uint32_t number);
 
 QString classString(uint8_t classVal);
-QString spell_name (uint16_t spellId);
+QString spell_name (uint32_t spellId);
 QString language_name (uint8_t langId);
 QString skill_name (uint16_t skillId);
 QString size_name (uint8_t size);
